FIX: #13 Autobox arguments to the ksonnet-lib constructors
This commit addresses two bugs. First, passing a container object into
`container.default` causes us to emit a single `v1.Container` object
instead of an array. Second, "special" parameters like `kind` and
`apiVersion` are sometimes exposed to the user, when they should be
automatically inferred from the library.

The second is trivially addressable.

The first is slightly more subtle. Technically, is spec-compliant: any
constructor that expects a parameter to be an array will expect the user
pass an array in as an argument. But, this is unintuitive behavior for
most users, who probably expect it to automatically box parameters that
are just a single value.

A better option (which we do thorughout the codebase) is to insert a
runtime conditional that checks whether the value passed in is an array,
and if it isn't, to box it as an array. In the above example, if the
user passed a raw container object, this would result in an invalid
Kubernetes API object, since the specification expects an array; with
this change, the constructor would package that up in an array on behalf
of the user.