Handle types more intelligently in property methods
Currently in a property method, we'll always emit a mixin, regardless of
whether it makes sense for the property type. For example:

  deploymentRollback:: {
    name(name):: {name+: name},
  }

In this case, `name` is a string, so using the `+:` operator is not what
we want.

This commit will change `propertyMethod#emit` to fix this problem
individually for each type in the OpenAPI spec.

Signed-off-by: Alex Clemmer <alex@heptio.com>