Add property blacklisting, identifier rewriting
When we emit `ksonnet-lib`, it is very useful to have the ability to
elide properties that the client should not be using (such as the
`status` field in `io.k8s.kubernetes.pkg.apis.apps.v1beta1.Deployment`),
and to rename properties that are either problematic (such as any
property whose name is a Jsonnet identifier) or not Jsonnet-style (such
as `podIP`).

This commit will implement this behavior. Specifically, such
transformations are specified by altering `kubeversion/data.go`, which
on a per-Kubernetes-version basis specifies:

1. Which properties of any fully-qualified API object name to blacklist
   (i.e., ignore when emitting Jsonnet code).
2. Transformations for the name of any property method (e.g., changing
   the name of, say, `TokenReviewStatus.error` ->
   `tokenReviewStatus.checkError`) to avoid collisions with Jsonnet
   keywords, and to fix casing problems that are not idiomatic to
   Jsonnet.

In particular, this means that updating `ksonnet-lib` for a new version
of Kubernetes involves simply adding rules for that version in
`kubeversion/data.go`.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>