Remove `propertyMethod`'s dependency on `kubespec.Property`
Currently `propertyMethod` depends on `kubespec.Property`. `Property` is
somewhat opaque, and none of the `kubespec` data structures have real
constructors, which will be somewhat inconvenient when we decide to
allow property aliasing, which will involve constructing a new
`propertyMethod` when we perform the mapping `apiObject` ->
`propertyMethod`.

This commit will pull out the direct dependency on `kubespec.Property`
and add fields for the properties we do care about.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>