Add type aliases
Currently, non-top-level API objects do not get emitted at all in
`ksonnet-lib`. This is particularly bad if a user wants to use the
normal abstractions of `ksonnet-lib` to construct (say) a
`v1.Container`, and pass that into a `Deployment` later.

This commit introduces inline type aliases. For example,
`deployment.spec` would be accompanied by a `deployment.specType`, which
is simply an alias for the type `DeploymentSpec`.

This carries 3 basic advantages:

1. Even experienced Kubernetes engineers tend not to know all the names
   of these types, since they are usually embedded inside top-level API
   objects (as `DeploymentSpec` is inside `Deployment`).
2. The huge number of non-top-level API objects overwhelms autocomplete.
3. Embedding the type aliases where they are used makes it much easier
   to access them, since you don't have to look up the object type.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>