Emit non-top-level API objects in a `local`
This commit will add all non-top-level API objects to a `local` that is
not visible to autocomplete or consumers of `ksonnet-lib`, in
preparation for the incorporation of type aliases.

For context, currently, we do not emit any non-top-level API objects as
part of `ksonnet-lib`. The reason is that adding them mainly serves to
confuse users: even experienced Kubernetes developers do not know these
object types by name (they are often anonymously embedded inside
top-level API objects, e.g., `DeploymentSpec`), and when they are all
emitted, they overwhelm autocomplete suggestions.

Still, users need to be able to write configurations that use these
types, so here we prepare to introduce them through embedded type
aliases (e.g., `deployment.deploymentSpecType`) by placing them in a
`local` called `hidden`, which is available only internally (i.e., not
visible to autocomplete suggestions or consumers of the library).

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>