Rename `propertyMethod` -> `property`
`propertyMethod` currently is used only to represent methods that set
properties in the Kubernetes API objects (e.g., `container.image`, which
sets `v1.Container`'s `image` property).

This will not be true when we begin to use it to represent type aliases,
which will be important for non-top-level API objects, which currently
are not emitted at all.

To better suit the semantics we seek, this commit will rename
`propertyMethod` -> `property`.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>