Standardize on `log.Panic` for internal errors
As we continue to extend this tool to account for future versions of the
OpenAPI spec, we expect that many of the assumptions we are presently
making will not hold. In these situations, it is beneficial to fail with
a stack trace, rather than to simply fail with an error message.

This commit hence adopts `log.Panic` as the default failure, as it will
print both a message and a stack trace.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>