Add jq script to find `status` and `ListMeta` blacklist properties
Our first cut at creating blacklisted properties was to eyeball the
documentation and write down everything that seemed server-generated.

This commit will introduce the first principled attempt to collect these
things automatically: a jq script that will output an object whose keys
are fully-qualified API object names, and whose values are lists of
properties to blacklist. For example:

  {
    "io.k8s.apimachinery.pkg.apis.meta.v1.Status": [
      "status", "metadata"
    ]
  }

This would indicate that the fields `status` and `metadata` are to be
blacklisted from the `meta.v1.Status` API object.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>