FIX: #32 Append to, instead of replacing, arrays in property methods
In property methods that take arrays as arguments, (e.g.,
`v1.PodSpec.containers`) we allow both an array and a single object to
be passed in. The desired behavior is that we will "box" the single
object on behalf of the user. For example, if a user passes in an array of
container objects, we append that to the existing array; if they pass a
single container object, we will box that in an array and append it to
the existing array of containers.

As described in bug #32, passing in a singleton object will actually
cause the original array (which we are supposed to append to) to be
completely overwritten.

This commit will introduce a trivial fix.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>