Fix bug causing type aliases to not use rewrite rules
ksonnet-gen defines a set of mechanisms to perform arbitrary
transformations on identifiers found in the Kubernetes API spec. For
example, we might want to rewrite `scaleIO` (which is a Go-style
identifier) to `scaleIo` (which is a Jsonnet-style identifier).

We would expect that type aliases (which are automatically generated for
`$ref` types) also get rewritten, so that something like `scaleIO`
generates a type alias called `scaleIoType`, rather than `scaleIOType`.
Unfortunately, this is not currently the case.

This review will correct this bug.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>