FIX: #33 Introduce rewrite rules for types
As mentioned in the previous commit message, ksonnet-gen allows users to
perform arbitrary transformations to identifiers found in the Kubernetes
API spec. For example, users might want a Go-style identifier
`HTTPHeader` to be emitted in Jsonnet code as a more
Jsonnet-style-compliant identifier like `httpHeader`.

Unfortunately, when we introduced the `hidden` member, we forgot to
write rules for the new types we exposed to the user.

This commit will introduce rules to rewrite all the problematic
identifiers in ksonnet-lib.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>