FIX: #31 Add properties of type `IntOrString` to ksonnet-lib
Kubernetes has a notion of a property being either an int or a string.
This is true of, for example, some fields handling ports, which might be
either the name of a port or a number of a port.

Since the OpenAPI spec does not directly support union types, this is
hacked in as a `$ref`, which is called `IntOrString`.

ksonnet-lib is currently not handling this case correctly. The desired
behavior is to simply treat these properties exactly as we do properties
that are int or string (they are identical as far as codegen goes).
Instead, we emit nothing at all.

This commit will introduce code that correctly handles all properties of
this type.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>