FIX: #44 Group name emitted incorrectly
Quoting from issue #44:

> `apiVersion` usually takes the form `group/version`. In the future,
> `group` will be a domain, rather than a single word (as old groups
> are).

So, for example, the group of
`rbac.authorization.k8s.io.v1alpha1.ClusterRole` is actually
`rbac.authorization.k8s.io`, not `rbac`.

This is emitted as part of the OpenAPI spec, in `SchemaDefinition`'s
`x-kubernetes-group-version-kind` field. This commit will use this (if
available) as the group name when setting the `apiVersion` field.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>