Add 'self' helper to all property methods
Currently the `+` pattern is used prodigiously by ksonnet-lib. For
example, we might build up a container object with something like:

    container.new("nginx-depl", "nginx") +
    container.ports(port.new(80))

This commit introduces new "helper" construct that greatly cuts down on
verbosity, so that something like the following is possible:

  container
    .new("nginx-depl", "nginx")
    .withPorts(port.new(80))

This is supported, essentially, by having every method return `self`.
For example, in the case of `metadata.name`, we would have something
like the example below.

  name(name):: self + __metadataMixin({name: name}),

The essential nature of this change is to include the surrounding object
in the return, so that all the methods of (e.g.) `metadata` are
available on it.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>